+++ packages/awino-ui/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add info component to landing page
 - Add benefit component to landing page
 - Add guide component to landing page
+- Add asset component to landing page
 
 ## 03.03.2022